% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{buy}
\alias{buy}
\title{Click the earliest available time}
\usage{
buy(n_tries = 100, timeout_after = 30 * 60, end_earlier = TRUE, sleep_time = 1)
}
\arguments{
\item{n_tries}{Number of times to try refreshing before giving up.}

\item{timeout_after}{Number of seconds to try refreshing before giving up.}

\item{end_earlier}{If true, the loop will end after either \code{n_tries} is reached or \code{timeout_after} is reached. If false, both conditions must be met.}

\item{sleep_time}{Average number of seconds to sleep in between refreshes. Exact sleep time is randomly chosen from a uniform distribution of \code{sleep_time - sleep_time/2} to \code{sleep_time + sleep_time/2}.}
}
\value{

}
\description{
If no times are available, keep refreshing.
}
\examples{
\dontrun{
buy(sleep_time = 2)
}
}
