% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{check_out}
\alias{check_out}
\title{Check out}
\usage{
check_out(cart = "amazon fresh")
}
\arguments{
\item{cart}{One of \code{"whole foods"} or \code{"amazon fresh"}.}
}
\value{
A session object
}
\description{
Check out
}
\examples{
\dontrun{
check_out("amazon fresh")
}
}
