% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_chromedriver_version}
\alias{get_chromedriver_version}
\title{Get the version of Chromedriver that matches your installed version of the Chrome application}
\usage{
get_chromedriver_version()
}
\value{
A character vector
}
\description{
Currently only supports MacOS.
}
\examples{
\dontrun{
get_chromedriver_version()
}
}
