% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{get_in}
\alias{get_in}
\title{Log in to Amazon}
\usage{
get_in()
}
\value{
Session object is stored in the global environment.
}
\description{
Launch a Selenium session in a Chrome browser, attach cookies to it to bypass 2FA, and log in.
}
\examples{
\dontrun{
get_in()
}
}
