% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{run}
\alias{run}
\title{Do the thing}
\usage{
run(cart = "amazon fresh", ...)
}
\arguments{
\item{cart}{One of \code{"whole foods"} or \code{"amazon fresh"}.}

\item{...}{Further args to \code{\link{buy}}.}
}
\value{

}
\description{
Do the thing
}
\examples{
\dontrun{
run
}
}
