% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nom.R
\name{run}
\alias{run}
\title{Do the thing}
\usage{
run(
  cart = "amazon fresh",
  chromedriver_version = get_chromedriver_version(),
  ...
)
}
\arguments{
\item{cart}{One of \code{"whole foods"} or \code{"amazon fresh"}.}

\item{chromedriver_version}{String supplying the version of chromedriver to use. Currently \code{get_chromedriver_version}}

\item{...}{Further args to \code{\link{buy}}.}
}
\value{

}
\description{
Do the thing
}
\examples{
\dontrun{
run
}
}
